Chore: Vertex AI 임베딩을 수동 구성으로 전환
